requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
